Send QP ATTACH and DETACH events in the VMX.
Emit attach and detach events when a queue pair peered with the
hypervisor context ID is alloced or detached by the guest.

Now that I understand these events better I have restored the validity
checks and made the attach and detach events valid in the VMX. The
QP resumed event is still not relevant in the VMX so the vsock layer
should not subscribe to it.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>