HGFS: clean up server packet information duplication
Each packet sets a bool in the HgfsPacket object to state whether the
packet can be processed asynchronously by this transport.

This data is not required and doesn't make sense to duplicate this for
every packet received by the transport. The same information is sent to
the HGFS server at connect time from the transport. This information is
part of the transport characteristics and the server already makes use
of some of these flags, e.g. shared memory support for some operation
requests like change notification.

Replace the packet bool with the correct transport characteristics flag
check for asynchronous packet handling capabilities.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>