Fix VMCISock_GetAFValue() for upstream vsocket testing
The Tools driver exposes an IOCTL to get the address family, since
it's not fixed.  The upstream driver was given a fixed address family,
so it doesn't need this IOCTL.  But our tests (and other apps) still
try to use the IOCTL, which causes them to fail with the upstream
driver.  Fix our header (the Tools version) so that it can cope with
both.

Note: not very performant, so best to cache the result if possible.
The test suite common code already caches it.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>