HGFS: Clean up server of the reply packet handling
The reply packet handler was returning a bool when in fact it cannot
fail. Fix this up and make it consistent with other calls for the reply
packet handler when filling the header in the complete request common
function.

This is a lot of repeating the same change for all the opcodes of each
supported request and version.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>