Make SSD-based backing stores visible to guests
This change determines if the VM's virtual disk is on SSD and make this
information available to guest. This is already implemented on ESX and
Mac OS. This change implements it on Windows and Linux.

On Linux rotational property of the disk and block device corresponding
to logical volume can be read from /sys/block/dev-name/queue/rotational

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>