Checkpoint routines for AsyncSockets backed by VSockSocket.
These checkpoint routines work a little differently than those for
VSockSocket. While the vSocket infrastructure checkpoints all sockets
which are open (because they are in the global VSockSocketStreamGlobal
tables) each AsyncSocket must be checkpointed by its owner. This gives
the owner the opportunity to reinstall all the necessary callbacks on
resume.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>