Allow vSockets to the VMX to survive a context ID change.
On vMotion there is an (unlikely) possibility that the context ID of the
virtual machine will conflict with a VM that is already running on the
destination host. To prevent this from distrupting communication a few
changes are necessary.

- The entire VMCI queue pair handle must be checkpointed, allowing the
queue pair to still be used.

- The guest driver must assume that if the context ID it receives as
the destination of a VMCI datagram is not the same as the one it
was expecting then its context ID must have changed and it should
update it. This assumption is valid because the hypervisor will never
deliver a datagram with the wrong context ID.

- To make the new test-vmx test work the range check in VMCI.SetID
had to be removed. Vigor doesn't support unsigned integers so a
negative value for the new context ID is valid.

A new version of the ttylinux iso is included containing a proposed
version of this patch for upstreaming.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>