HGFS: clean up the transport channel callbacks from the HGFS server
The transport channel callbacks for the HGFS packet request-reply to read,
write and release the memory use badly defined arguments. It uses character
pointer and calls it a token. The type is simply just wrong and should be
treated as an opaque type owned and manipulated only by the channel transport.
Secondly, it is really a context although it is a token, it could be anything
the channel chooses and could change in the future.

This change fixes these issues to be a void *context. This shows that the type
and meaning are not to be exposed beyond the owner of object which is the channel
in this case.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>