OVT: require glib2 >= 2.14 and disable 'deprecated-declarations' warnings
Internally we are using 2.14.2 on FreeBSD, 2.16.4 on Solaris, and
2.24.2 on Linux, so let's stop pretending that we work with earlier
versions. RHEL5 with 2.12 hasn't been compiling for a while and even
if it would it would have reduced functionality as glib_regexp would
not be available. FreeBSD started packaging glib2 2.22 starting with 7.1.

Also, mutex API in Glib2 is a mess, they keep changing and deprecating
it, so for now simply add -Wnoerror=deprecated-declarations to avoid
build errors.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>