HGFS: Improve VMCI transport headers part I
Begin to consolidate the client to server and server to client VMCI transport
headers to be just one common header for both.
Currently, there are two one for each direction and they are almost
identical. Also the header for the client to the server does not
have all information that is required for asynchronous packet sending
and replies. Hence the motivation for improving this.

The packet headers are poorly defined as they do not have any reserved
fields for extending the current version and they are inconsistent with
the packet types with each direction.

This initial change pulls out the version and packet type into a header
node that will be common to the version one and newer versions of the
transport header.

This involves modifying the server and client sides. The version 1 transport
header will remain binary compatible with this change, it just moves the
first two fields into a common substructure.

Broke out the transport header to return the information from multiple
versions of the header.
Moved the validate datagram function call to the main receive message callback.
Removed the transport header check from this function too, as it is not
useful to have it there. It will be done by each of the specific header version
handling routines.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>