HGFS: Fix Posix server from blocking on pipes
The HGFS protocol function to get attributes of a file or folder does
a Posix open without the O_NONBLOCK flag. This can cause the open to block
if there isn't anyone on the other end of the pipe.

Fix is to use the same flags we use for the protocol open to ensure the
correct basic flags are used. Also, if the share is not allowed to follow
symlinks we get this flag set correctly too now.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>