HGFS: Fix notification cleanup session tear down
Previously, when a transport session was disconnected any notifications
for the HGFS session were then removed if notification module was enabled.

However, this is grossly inadequate, as an HGFS session can come and go driven by
a client's protocol requests. The notifications are per HGFS session which is
completely independent from the transport session. The only reason the transport
session comes into play is whether it has the transport characteristics to support
the bidirectional nature of the notification feature.
When an HGFS session is destroyed whether it is from a tranport connection
session disconnect or a destroy HGFS session protocol request any folder
the outstanding notifications for the session should be removed.
Since that currently does not happen, e.g. when an HGFS session is destroyed via the
protocol requests from a client and goes on to create a new one, during
a Windows VM reboot, then when host updates do occur on an HGFS share the
notifications will be generated that hold pointers to invalid sessions.
This causes the VMX to crash.

The fix is simply to move the notification teardown for a session is at the session
teardown function, from the transport session disconnect.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>