Add compiler barriers to Atomic_MFence
The MFENCE instruction ensures that all loads and stores that
preceed it are "globally visible" before any load or store that
follow it.  However, we don't have a compiler barrier in
Atomic_MFence and so the compiler can move code and defeat this.

While this bug has probably been around for quite a while, it
only came up recently now that the AtomicEpilogue() is compiled
out.  That contains a compiler memory barrier, and while it's
not technically sufficient (need barrier before the MFENCE too),
it was apparently enough to discourage the code motion in
practice.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>