HGFS: Clean up the VMCI transport request and reply headers part I
This is the first part of cleaning up the HGFS VMCI transport request and
reply headers. This deals with the reply header only which is almost identical
to the client request VMCI header but not quite - frustratingly.
It really is not ideal to use a reply header that is different from
the request header and has fields that require corresponding information from
the request header which are missing. Consequently, the request header is
going to be ramped up to a version 2 which includes all the information for both
request and reply and be extensible.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>