HGFS: Fix Linux client to work with Perforce versioning
The problem is that perforce client uses rename operation upon a temp
file which has read-only attributes. The operation then fails with
permission denied.

The fix removes the read only attribute and retries the rename operation
again as per the delete operation. This occurs on Windows HGFS servers
as the target of the rename operation to be replaced has the read only
attribute set. The prevents a rename even if they want to overwrite the
target if it exists. So removing the read only flag is required.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>