HGFS: Fix the server to close sessions being invalidated
When the HGFS server running in the tools is left with any sessions
open due to the clients not terminating them cleanly they will be marked
inactive. The HGFS server callback to invalidate them will periodically
be called and after a brief period will terminate those sessions.
These sessions are not closed first, thus triggering the assert in the
session exit function which checks the session's state.

Fix is to close the inactive sessions being terminated in the HGFS
server invalidate callback.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>