HGFS: Clean up HGFS server packet abstraction part II
Clean up the input params object which reuses the packet abstraction
field names which are generic because they are opaque outside of the HGFS
server. Since the input params are wholly contained within the HGFS server
and have specific meaning in the HGFS server context, we give the fields
the HGFS meaningful names.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>