HGFS: Clean up server packet abstraction part IV
Complete some tidy up of the input params created from the HgfsPacket.
This can be contained to only the HgfsServer code and not required to be
exposed beyond that. So this removes the one usage from the parameter
pack and unpack code which was in the unpack write request.

As I was modifying up the HGFS server write request I correct a couple of
const char * to const void * for the write data.

I removed a HSPU_PutPacket declaration as the function did not exist.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>