HGFS: Clean up server transport session usage
The transport session is only ever used in the common HGFS server code.
The scope should be constrained to that file and beyond that treated as
an opaque type.

Make some of the transport session functions static to the common HGFS
server file and remove the declarations from the common header file.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>