Avoid a memory dereference when getting the current thread ID.
As part of some recent benchmarking and ensuing discussions a couple
of overheads in VThread_CurID have been identified.

(1) VThread_CurID ultimately gets a pointer to a per thread
structure that contains various bits of state including the
thread ID.  After getting that pointer it must read the thread
ID out of memory.  This could be avoided if the thread local
value we stored was the thread ID itself.

(2) Before calling into the host's API for thread local storage,
vthreadBase.c must get the host key, check if it is initialized
and initialize it if necessary.  Then once it gets the thread
local value from the host API it must check if that has been
initialized.  These checks could be avoided if we forced clients
to initialize the thread before calling functions like
VThread_CurID.

(3) On some (all?) Linux pthread implementations pthread_getspecific
itself can be heavyweight.  Both Windows and Linux offer
alternatives to make thread local storage cheaper.  (OS X on the
other hand provides a very fast implementation of
pthread_getspecific: pretty much one instruction plus the
function call overhead.)

The first two overheads came up in my profiling of USB workloads on OS
X while Kevin raised the third issue in the following discussions.

This change attempts to eliminate the first overhead, but to do so in
a way that helps set up the code for attacking the remaining two.  In
particular it introduces a second thread local variable to store the
thread ID.  For the time I left the thread ID in the other structure
as well and verify that they stay in sync.  We could get rid of it,
but it's low cost and I suspect (though I have no proof) that it could
be useful in debugging.

As part of this I've rearranged some of the initialization code both
for the TLS keys as well as for initializing the TLS data.  This is
useful because I wanted both pieces of state to get set together (and
mutated together -- yes our threads change IDs during their life).
And, with respect to the pthread keys, it's important to make sure the
base key gets allocated before the thread ID key.  This ensures it
gets destroyed first which allows us to keep an ASSERT to make sure
that the two thread IDs stay in sync.

This does run into one wrinkle because of the lazy thread
initialization.  The default value for uninitialized state is NULL
which, inconveniently, is a valid vthread ID.  So instead of actually
storing the thread ID, we store the thread ID incremented by one.
This means the default/uninitialized value will show up as -1.  We can
avoid this after addressing (2) from above, but in the meantime
trading a memory read for an ALU operation is still a nice win.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>