Flip the kill-switch on VM2VM VMCI
VM2VM VMCI made a reappearance in 5.5 because vHadoop was planning to
use it, but it was eventually rejected.  We left the code in so as not
to disturb stability, and simply disabled it in the UI.  But let's
make doubly sure by flipping the kill-switch, so that our customers
don't accidentally enable it and then complain about it being removed
from 5.1, coming back in 5.5 and then disappearing again.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>