VMCI: Remove VM2VM
We're never going to ship it, so let's yank it out.  This isn't a
straight backout for the following reasons:

o Some of the files have diverged so far that backout fails.
o Petr added some nice log messages in his original change that we
want to retain.

So this was mostly manual.  I did do a side-by-side with the pre-VM2VM
version and it looks good.

Note that our 128-bit QP tag code will re-appear in vm_atomic.h at
some point.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>