HGFS: Server packet abstraction part IX
The HGFS server sets the packet's ID field as the HGFS protocol request
ID.  This should not occur and it overloads that the packet ID which
should be private to the transport channel which has nothing to do with
the HGFS request ID.  The HGFS request ID has it's own field in the HGFS
header used by the request and reply.

Remove the HGFS server modifying the packet ID.  This is currently safe,
as the clients do not receive any asynchronous replies from any HGFS
server where the packet ID in the reply is used.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>