Use __thread on Linux (and ESX) in vthreadBase.c.
This change makes use of __thread on Linux for storing the vthread
base pointer and the thread ID.  We still also store the value with
pthreads_setspecific so that we have consistent behavior on thread
destruction (using the pthreads destructor).  But this allows the
VThread_CurID path to skip loading the pthread key, skip checking if
the key is initialized, and skip the call into pthread_getspecific.

(With VThread_CurID taking its own fast path, it's not clear there's
much benefit at all from optimizing the path for the base pointer, but
I did so for consistency's sake.)

I tried using my USB bandwidth test to measure the performance impact
of this on a Linux host.  Unfortunately I found the test to be much
more noisy under Linux.  The results claimed 1.6% faster overall, but
I have almost no faith in that number.

Still, having inspected the new code and looked at the contents of
pthread_getspecific, I think it's safe to claim that this should be
better overall even if its difficult to quantify how much.  And though
it makes the code slightly more crufty, such ugliness is confined to
two functions.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>