HGFS: Server packet abstraction part IX
Adding in the new fields to indicate the size of the valid data, not the
size of the buffer. Previously, the two were combined which meant that
the buffer size was ultimately used instead making the data copies very
inefficient.

Another  transitional change to get this stuff moved over to handle the
total pack size and valid data sizes separately. Currently, VMCI has to
use the total packet sizes only and so maps in 16 pages and copies all
the data back and forth because it can't tell how much is actually
valid.  When I switch the VMCI transport header to support V2 it will
have the correct data size and packet ID - asynchronous will then work
too.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>