Switch utf::string from g_utf8_validate to Unicode_IsBufferValid.
In the version of glib we currently use g_utf8_validate rejects strings
containing the non-characters U+FFFE and U+FFFF. This was marked as a
bug and resolved (see https://bugzilla.gnome.org/show_bug.cgi?id=694669)
but until that is released we need a workaround. Our internal UTF-8
validation routine accepts these characters as valid so use that
instead.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>