MPN to MPN 64 conversion for Hosted.
This change is supporting a 64bit MPN for hosted products.

Assumptions taken:

1. Kindly note that currently we do not have systems supporting 64bit
memory. So hosted vmmon 64 bit MPN is a container to store whatever
monitor passes to it.  The underlying assumption is that the higher
32bits will be zero.

2. Physmem tracker still does page walk using 32 bit values and we have
put checks to make sure that MPN64 is just a container with actually 32
bit values. This was a design decision taken after talking to kevinc and
jpool.

3. The CrossPage code is still 32 bit and I typecast MPN64 to accept 32
bit values for the crosspage code.

4. Vmkernel already has another definition of MPN64 and I have retained
their definition for vmkernel specific code.

typedef uint8 * MPN64;
typedef uint64 MPN64;

5. Additional changes were required in bora-vmsoft vmballoon module
which earlier used MPN and called functions expecting MPN64 as an input.
I have however *NOT* changed the PageHandle code of Vmballoon.  I have
typecasted to MPN64 where common MacOS page handling functions which
require MPN64 are used.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>