HGFS: Clean up server send packet callback
While cleaning up the usage of the HGFS packet abstraction it meant that
the arguments to the HGFS server packet send callback could be cleaned up.
The arguments for the packet buffer and size are now unused as this too
arguments are extracted from the HGFS packet object.
Therefore, remove the unused arguments from the send callback.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>