HGFS: Fix a Linux client memory corruption
When doing partial page writes the Linux HGFS client would trash the
initial part of the page leading up to the start of the write and also
trash the remaining part of the page after the written segment of the
page.

This trashing of the initial part of the page was incorrect and
overwrites earlier partial page writes to that page.  This was also
incorrect as the arguments for the offsets into the page were
miscalculated.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>