HGFS: Fix Linux client inode 512B block count
The Linux HGFS client miscalculated the number of 512 byte blocks to set
in the inode for the file size. It was incorrectly using the HGFS
hardcoded block size instead of 512, which was 1024. This causes the
block count to be half the correct number and so du on a file reports
the incorrect size.  fstat also retrieves this block count in the stat
structure so it is likely more applications will break.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>