VMCI: make guest driver work with virtual IOMMU
Right now we use vmalloc()/alloc_page() for our guest queuepair pages
(and bitmap page) and then do a virt_to_phys() before passing them down
to the device.  That's not going to work if DMA remapping is enabled,
since the IOMMU has no idea about the mappings.  Switch to
pci_alloc_consistent() instead.  We still allocate each page
individually, since there's no guarantee that we'll get a contiguous
block of physical for an entire queuepair (especially since we allow up
to 128 MiB!).

Also made split between guest and host in the kernelIf struct much
clearer.  Now it's obvious which fields are which.

Finally, drop the PINNED flag from Linux.  That was only ever used by
vVol-Filter, and that's been moved over vPageChannel, which doesn't use
queuepairs but has custom rings.

Once this is in I'll port it to upstream.  I considered doing upstream
only, but that would mean pre-3.9 guests would be unable to use
IOMMU support.  We should really freeze the in-house driver after
this...

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>