Fix AssertOnCompileFailed for GCC 4.8 (kernel 3.10+)
GCC 4.8 now includes -Wunused-local-typedefs as part of -Wall.  This
causes it to spew a gazillion warnings when compiling our modules on
3.10+ kernels.  Because it's a typedef we can't use the (void) cast
trick, so we need to use attributes.  This change introduces
UNUSED_TYPE to vm_basic_types.h and uses it where appropriate to make
the compiler happy when building our modules.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>