Implement VMware Tools for ESXi 6.0 guests.
This change provides limited VMware Tools support for ESXi 6.0
guests.  The powerOps and guestInfo plugins are supported.
Only IPv4 addresses are reported, since those are the only
ones we can get through the vmkuser library.

VMware Tools for ESXi 6.0 are packed as two VIBs on a single
ISO image.  One VIB contains the glib libraries that are
missing from the standard ESX distribution, and the other
VIB contains the tools themselves.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>