Fix AsyncSocket reference leak when using IVmdbPoll
A reference is taken when an AsyncSocket callback is registered in
IVmdbPoll to protect the AsyncSocket from being freed while the callback
has been scheduled to run.  That reference is released when the callback
is unregistered if the callback is not going to run, or from the
callback itself if it is already scheduled.  The current code does not
correctly handle the case when the callback unregister itself, as it
needs to explicitly release the reference in that case.  This change
also adds a Bool to AsyncSocket so we can distinguish between send
callback that is registered as a timer callback so that we know which
type of callback to remove, which is necessary to keep the reference
count correct.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>