Fix the build on FreeBSD 12.0-CURRENT.
On FreeBSD 12.0-CURRENT we've extended the sysctl_add_oid() function to
include an additional argument: an aggregation label. This doesn't tend
to be problematic, because (almost) all users of these interfaces use of
the recommended macro SYSCTL_ADD_OID(), which didn't get modified.

Fix up this piece of code to make use of this macro as well, so that it
builds on all (supported) versions of FreeBSD.