Fix timesync state to be uncalibrated when adjustment is large (>60ms).
Open-vm-tools has a timesync plugin to adjust the guest time to match the
host time. Timesync relies on a state machine (uncalibrated, calibrating
and PLL states) to adjust the amount of guest time to match host time.
If the guest and host time differ by more than 60ms, the state machine
should reset to the uncalibrated state. This change fixes a bug that could
leave the state in PLL mode even when the guest and host times differed by
more than 60ms, since the absolute value of adjustment was not considered,
which in turn caused the time sync to happen more slowly than expected.