VMCI: reset datagram buffer pointer on dealloc
Fix a potential guest crash when unloading the VMCI
driver while upgrading to a specific version of UEK.
A BH may run during module unload and attempt to deliver
datagrams to the datagram buffer after it has been
deallocated. It turns out that the tasklet may not
yet have been scheduled, so add a tasklet_disable to
guard against that.  Also, reset the datagram buffer
pointer when the buffer is deallocated to indicate
that it is no longer there.