Add a stolen time interface for guests
This change adds a cheap stolen time interface for guests. Guests must
use the backdoor port to register a per-VCPU object in memory called the
"stealclock" with the monitor. From then on, until the clock is
disabled, the monitor publishes clock values whenever stolen time is
accumulated. A complete specification of the interface is included in
this change, as a text document: bora/doc/stealclock.txt.

Pending clock updates are represented by an idempotent monitor action,
posted whenever stolen time is accumulated. In the current
implementation, this is whenever we update stolen time on transit from
host to vmm. In the future, it may an action posted by the vmkernel,
whenever stolen time is updated.