Tools: Implement resolutionKMS for linux
tools: Enable the use of the PKG_CHECK_MODULES macro for open-vm-tools

Adds a path to the pkg-config utility and sets up the open-vm-tools test
build environment to be aware of its m4 macros.
Also sets up a local directory for .pc files and adds the libdrm and libudev
pc files to that directory (during open-vm-tools configure time) to prepare
for resolutionKMS testing.

tools: Prepare resolutionSet for resolutionKMS on Linux

- Add a set of utilities for detecting the DRM version, the Xorg driver
version and to dlopen libudev and libdrm if they are not available at
build time. The utilities also include opening and closing a DRM file
descriptor from a given type of device node.
- Change the check for Xorg driver version in resolutionX11 to use the
new utility
- Pass a ToolsAppCtx to the resolutionSet backend for access to the
configuration database.
- Finally, detect if there is system and / or configuration support for
resolutionKMS and in that case back off resolutionSet

- Added some license information to resolutionDL.h
- Added an include to resolutionDL.h
- Fixed some ifdefs around HAVE_RESOLUTIONKMS
- Modified other backends for the new resolutionToolkitInit() interface.

tools: Add the resolutionKMS plugin

What's resolutionKMS
ResolutionKMS is a plugin that takes over the resolutionSet functionality,
but as a vmsvc plugin rather than a vmusr plugin. Instead of talking to the
X server it talks to the Kernel Mode Setting (KMS) functionality of the
vmwgfx kernel module. This enables the following functionality.

- Resolution / Topology switching at the login screen
- Resolution / Topology switching when multiple users have launched multiple
X servers as part of fast user-switching.
- Resolution / Topology switching for new native compositors that support
the KMS way of communicating resolution / topology (as preferred mode and
connector properties) As far as I know, only gnome-shell / Wayland can do the
topology switching today, since I implemented the missing support for it.
We need to address this issue also with Mir and other Wayland implementations.

New dependencies.
The new code requires libudev and libdrm. For open-vm-tools, I've opted to
require those libraries at configure time. There's no option to build linux
open-vm-tools without resolutionKMS support, although that should be easy to
add if needed. For vmware-tools, there is code to enable resolutionKMS iff
libudev.so.1 and libdrm.so.2 are present at load time and can be dlopened.
For vmware tools there are no new build dependencies introduced at load time
AFAIK. (dlopen?)

Enablement
ResolutionKMS requires vmwgfx 2.10.0+ and xf86-video-vmware 13.2.0+ to
guarantee that we don't lose functionality compared to old resolutionSet. The
vmwgfx kernel module has a way to check version at runtime. For the xorg driver
the situation is different since resolutionKMS is typically launched before
Xorg. Therefore we fall back to the hackish way previously implemented in
resolutionX11.c to find the driver file and scan it for version info. Oh well.
In any case, there is also an option to override this and enable or disable
resolutionKMS in the tools.conf file. If resolutionKMS is disable either by
lacking system support or by configuration, the old resolutionSet functionality
kicks in. Extensive debugging info will be present in the tools daemon logs.